Updated signature to match

+++ sbsur/sbs.pxd
@@ -5,4 +5,4 @@ from sequence_generator cimport SequenceGenerator
 
 # We hand out a vector (instead of an array) of sequences so that it's automatically freed
 # Note that the sequences that we output are in reversed order
-cdef vector[(vector[int], float)] c_sample(SequenceGenerator generator, int batch_size)
+cdef vector[(vector[int], double)] c_sample(SequenceGenerator generator, int batch_size)
